docker/camera-agent/requirements.txt
opencv-python-headless>=4.8
websockets>=12.0
aiohttp>=3.9
zeroconf>=0.130
PyYAML>=6.0